add timeout expected/unexpected
some iterations of the shell cannot `exit,` and some incorrect
implementations may get stuck in an infinite loop.

To combat this, the test program only runs the shell for a very
short time, and takes not of when the program does/does not run into
this maximum run duration timeout.

Tests should be marked as expected to timeout in tests.json